# Optimize simple range mappings: `[for n in start..finish -> f n]`, &c.