--- a/src/Compiler/Optimize/LowerComputedCollections.fsi
+++ b/src/Compiler/Optimize/LowerComputedCollections.fsi
@@ -2,9 +2,15 @@
 
 module internal FSharp.Compiler.LowerComputedCollectionExpressions
 
+open FSharp.Compiler.AbstractIL.IL
 open FSharp.Compiler.Import
 open FSharp.Compiler.TcGlobals
 open FSharp.Compiler.TypedTree
 
 val LowerComputedListOrArrayExpr:
-    tcVal: ConstraintSolver.TcValF -> g: TcGlobals -> amap: ImportMap -> Expr -> Expr option
+    tcVal: ConstraintSolver.TcValF ->
+    g: TcGlobals ->
+    amap: ImportMap ->
+    ilTyForTy: (TType -> ILType) ->
+    overallExpr: Expr ->
+        Expr option